# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "FeishuConfigUpsertRequest": {
        "properties": {
          "alert_enabled": {
            "default": true,
            "title": "Alert Enabled",
            "type": "boolean"
          },
          "analysis_enabled": {
            "default": true,
            "title": "Analysis Enabled",
            "type": "boolean"
          },
          "app_id": {
            "title": "App Id",
            "type": "string"
          },
          "app_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Secret"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "news_batch_interval_minutes": {
            "default": 60,
            "title": "News Batch Interval Minutes",
            "type": "integer"
          },
          "news_enabled": {
            "default": true,
            "title": "News Enabled",
            "type": "boolean"
          },
          "news_keywords": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Keywords"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "default": "chat",
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "app_id",
          "target_id"
        ],
        "title": "FeishuConfigUpsertRequest",
        "type": "object"
      },
      "FeishuConfigView": {
        "properties": {
          "alert_enabled": {
            "default": true,
            "title": "Alert Enabled",
            "type": "boolean"
          },
          "analysis_enabled": {
            "default": true,
            "title": "Analysis Enabled",
            "type": "boolean"
          },
          "app_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Id"
          },
          "app_secret_set": {
            "default": false,
            "title": "App Secret Set",
            "type": "boolean"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "news_batch_interval_minutes": {
            "default": 60,
            "title": "News Batch Interval Minutes",
            "type": "integer"
          },
          "news_enabled": {
            "default": true,
            "title": "News Enabled",
            "type": "boolean"
          },
          "news_keywords": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Keywords"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "configured"
        ],
        "title": "FeishuConfigView",
        "type": "object"
      },
      "FeishuTestResult": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "FeishuTestResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "News Caught Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notify/feishu/config": {
      "get": {
        "operationId": "get_feishu_config_api_notify_feishu_config_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeishuConfigView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feishu Config",
        "tags": [
          "notify"
        ]
      },
      "post": {
        "operationId": "upsert_feishu_config_api_notify_feishu_config_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeishuConfigUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeishuConfigView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Feishu Config",
        "tags": [
          "notify"
        ]
      }
    },
    "/api/notify/feishu/test": {
      "post": {
        "operationId": "test_feishu_notification_api_notify_feishu_test_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeishuTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Feishu Notification",
        "tags": [
          "notify"
        ]
      }
    },
...
  }
}